requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
pluggy==1.5.0
pygments==2.18.0
pytest==8.2.1
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
rich==13.7.1
ruff==0.4.6
shellingham==1.5.4
six==1.16.0
spotipy==2.24.0
tomli==2.0.1
typer==0.12.3
typing-extensions==4.12.0
urllib3==2.2.1
